refactored to only handshake + boxstream
