--- code/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "code"
-version = "0.1.0"
-authors = ["Dhole <dhole@riseup.net>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rand = "0.6.0"
-
-ed25519-dalek = "1.0.0-pre.1"
-sodiumoxide = { git = "https://github.com/Dhole/sodiumoxidez", branch = "extra" }
-
-base64 = "0.10.1"
-hex = "0.3.2"
-get_if_addrs = "0.5.3"
-arrayref = "0.3.5"
-
-async-std = { version = "1.1.0", features=["unstable","attributes"] }
-
-crossbeam = "0.7.2"
-log = "0.4.8"
-env_logger = "0.7.0"
-
-serde = "1.0.102"
-serde_json = { version = "1.0.41", features=["preserve_order","arbitrary_precision"] }
-serde_derive = "1.0.102"
-
-dirs = "2.0"
-
-futures = "0.3.1"
-
-[dev-dependencies]
-test_utils = { path = "test_utils" }
