--- code/src/lib.rs
@@ -1,18 +1,0 @@
-#[macro_use]
-extern crate arrayref;
-extern crate log;
-extern crate serde;
-extern crate serde_json;
-#[macro_use]
-extern crate serde_derive;
-extern crate base64;
-
-#[macro_use]
-pub mod utils;
-
-pub mod boxstream;
-pub mod boxstream_sync;
-pub mod handshake;
-pub mod handshake_async;
-pub mod handshake_sync;
-pub mod pasync;
