--- code/src/pasync/crypto/mod.rs
@@ -1,3 +1,0 @@
-mod sodium;
-
-pub use sodium::ToSodiumObject;
