--- code/src/pasync/mod.rs
@@ -1,5 +1,0 @@
-pub mod handbox;
-pub mod rpc;
-pub mod util;
-pub mod crypto;
-pub mod patchwork;
