--- code/src/pasync/patchwork/mod.rs
@@ -1,16 +1,0 @@
-mod api;
-mod config;
-mod messagetypes;
-mod feed;
-mod message;
-mod encoding;
-mod privatebox;
-mod pubs;
-
-pub use api::{
-    parse_feed, parse_latest, parse_message, parse_whoami, ApiClient, CreateHistoryStreamArgs,
-    CreateStreamArgs,
-};
-pub use config::{ssb_net_id, IdentitySecret};
-pub use privatebox::{is_privatebox,privatebox_cipher,privatebox_decipher};
-pub use encoding::{ssb_sha256,stringify_json};
