--- code/src/pasync/rpc/mod.rs
@@ -1,3 +1,0 @@
-mod rpc;
-
-pub use rpc::{RpcClient,Header, RequestNo, RpcType};
