--- code/src/pasync/util/error.rs
@@ -1,7 +1,0 @@
-use async_std::io;
-use std::fmt::Debug;
-
-pub fn to_ioerr<T: Debug>(err: T) -> io::Error {
-    io::Error::new(io::ErrorKind::Other, format!("{:?}",err))
-}
-
