--- code/src/pasync/util/mod.rs
@@ -1,5 +1,0 @@
-mod error;
-mod buffer;
-
-pub use error::to_ioerr;
-pub use buffer::CircularBuffer;
