+++ examples/async.rs
