+++ examples/broadcast.rs
