+++ examples/ed25519_2.rs
