+++ src/async_std/circularbuffer.rs
@@ -344,5 +344,3 @@ mod test {
         Ok(())
     }
 }
-
-
