+++ src/async_std/mod.rs
@@ -1,5 +1,8 @@
 mod boxstream;
 mod handshake;
 
+#[cfg(test)]
+mod circularbuffer;
+
 pub use handshake::{handshake_client,handshake_server};
 pub use boxstream::{BoxStreamRead,BoxStreamWrite,BoxStream};
