+++ src/lib.rs
@@ -1,0 +1,18 @@
+#[macro_use]
+extern crate arrayref;
+extern crate log;
+
+#[macro_use]
+pub mod utils;
+
+#[cfg(feature = "sync")]
+pub mod sync;
+
+#[cfg(feature = "async_std")]
+pub mod async_std;
+
+mod boxstream;
+mod handshake;
+
+pub use boxstream::*;
+pub use handshake::*;
