+++ src/sync/mod.rs
@@ -1,0 +1,5 @@
+mod boxstream;
+mod handshake;
+
+pub use boxstream::BoxStream;
+pub use handshake::{Error, Result, handshake_client, handshake_server};
